write console application code for client

+++ PlaneShapesLibrary/Properties/CircleProperty.cs
@@ -1,7 +1,7 @@
 public class CircleProperty
 {
     public double Radius { get; set; }  
-    public CircleProperty(int radius)
+    public CircleProperty(double radius)
     {
         if (radius <= 0)
             throw new InvalidDataException("PropertyError: Radius must be positive number");
